+++ src/hooks/useIndexer/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./lib/useIndexer";
+export * from "./lib/types";
