+++ src/page/view/BubbleSort/lib/createReceipt.ts
@@ -1,56 +1,12 @@
-import { Receipt } from "../indexing/types";
+import { IndexerReceipt } from "@/hooks/useIndexer";
 import { Command } from "../model/command";
-import { InitializeArgs, State } from "../model/types";
+import { State } from "../model/store";
+import getCommands from "./getCommands";
 
-// TODO: Use i18n on messages
-export default function createReceipt(
-  args: InitializeArgs
-): Receipt<State, Command> {
-  const commands: Command[] = [];
-  const draft = [...args.array];
-
-  const compare = (i: number, j: number) => {
-    commands.push({
-      type: "compare",
-      message: `Assert if ${draft[i]} > ${draft[j]} (arr[${i}] > arr[${j}])`,
-      payload: {
-        indexA: i,
-        indexB: j,
-      },
-    });
-    return draft[i] - draft[j];
-  };
-
-  const swap = (i: number, j: number) => {
-    const temp = draft[i];
-    draft[i] = draft[j];
-    draft[j] = temp;
-    commands.push({
-      type: "swap",
-      message: `Swap item ${i} and ${j}`,
-      payload: {
-        indexA: i,
-        indexB: j,
-      },
-    });
-  };
-
-  for (let i = 0; i < draft.length - 1; i++) {
-    for (let j = 0; j < draft.length - 1 - i; j++) {
-      if (compare(j, j + 1) > 0) {
-        swap(j, j + 1);
-      }
-    }
-  }
-
-  commands.push({
-    type: "done",
-    message: "DONE!",
-    payload: {},
-  });
-
-  return {
-    commands,
-    initialState: { array: args.array },
-  };
-}
+export const createReceipt = (array: number[]) =>
+  ({
+    initialState: {
+      array,
+    },
+    commands: getCommands(array),
+  } satisfies IndexerReceipt<State, Command>);
