+++ src/page/view/BubbleSort/lib/getCommands.ts
@@ -1,0 +1,49 @@
+import { Command } from "../model/command";
+
+// TODO: Use i18n on messages
+export default function getCommands(array: number[]): Command[] {
+  const commands: Command[] = [];
+  const draft = [...array];
+
+  const compare = (i: number, j: number) => {
+    commands.push({
+      type: "compare",
+      message: `Assert if ${draft[i]} > ${draft[j]} (arr[${i}] > arr[${j}])`,
+      payload: {
+        indexA: i,
+        indexB: j,
+      },
+    });
+    return draft[i] - draft[j];
+  };
+
+  const swap = (i: number, j: number) => {
+    const temp = draft[i];
+    draft[i] = draft[j];
+    draft[j] = temp;
+    commands.push({
+      type: "swap",
+      message: `Swap item ${i} and ${j}`,
+      payload: {
+        indexA: i,
+        indexB: j,
+      },
+    });
+  };
+
+  for (let i = 0; i < draft.length - 1; i++) {
+    for (let j = 0; j < draft.length - 1 - i; j++) {
+      if (compare(j, j + 1) > 0) {
+        swap(j, j + 1);
+      }
+    }
+  }
+
+  commands.push({
+    type: "done",
+    message: "DONE!",
+    payload: {},
+  });
+
+  return commands;
+}
