+++ src/page/view/BubbleSort/model/store.ts
@@ -1,22 +1,19 @@
-import {
-  createEnvironmentStore,
-  EnvironmentData,
-} from "../indexing/environmentStore";
+import { createEnvironmentStore } from "@/hooks/useIndexer/lib/environmentStore";
 import { Command } from "./command";
-import { State } from "./types";
 
-const defaultData: EnvironmentData<State, Command> = {
-  // LATER: Refactor this out
-  receipt: {
-    initialState: {
-      array: [],
-    },
-    commands: [],
-  },
-  state: {
-    array: [],
-  },
-  step: 0,
+/**
+ * The state that gets changed by algorithm.
+ */
+export type State = {
+  array: number[];
+  // LATER: add i & j here
 };
 
-export const useEnvironmentStore = createEnvironmentStore(defaultData);
+const initialState: State = {
+  array: [],
+};
+
+export const useEnvironmentStore = createEnvironmentStore({
+  initialState,
+  commandType: {} as Command,
+});
