--- src/page/view/BubbleSort/model/types.ts
@@ -1,14 +1,0 @@
-/**
- * The state that gets changed by algorithm.
- */
-export type State = {
-  array: number[];
-  // LATER: add i & j here
-};
-
-/**
- * Argument for creating `Receipt`.
- */
-export type InitializeArgs = {
-  array: number[];
-};
